Week 22 audit entity

+++ src/main/java/com/piseth/java/school/phoneshop/phoneshop/entities/Model.java
@@ -1,7 +1,8 @@
 package com.piseth.java.school.phoneshop.phoneshop.entities;
 
-import jakarta.persistence.*;
 import lombok.Data;
+
+import javax.persistence.*;
 
 @Data
 @Entity
@@ -9,8 +10,10 @@ import lombok.Data;
 public class Model {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
+
     private String name;
+
     @ManyToOne()
     @JoinColumn(name = "brand_id")  // using Column will error
     private Brand brand;
